+++ ansible/modules/consul_acl/consul_acl_token.py
@@ -163,7 +163,7 @@ class ConsulAclToken(object):
 
     def update_token(self, accessor_id, **kwargs):
         current = self._make_api_request("acl/token/" + accessor_id, "GET")
-        if current["policies"] is None:
+        if current.get("policies") is None:
             current["policies"] = []
 
         if not has_token_changed(current, kwargs):