+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-01-24
+
 ### Added
 
 -   Add ability to create tokens with a specific accessor ID.
@@ -42,6 +44,7 @@
 -   `consul_acl_policy` module.
 -   `consul_acl_token` module.
 
-[unreleased]: https://github.com/jsundh/ansible-modules-consul-acl/compare/0.2.0...HEAD
+[unreleased]: https://github.com/jsundh/ansible-modules-consul-acl/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/jsundh/ansible-modules-consul-acl/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/jsundh/ansible-modules-consul-acl/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/jsundh/ansible-modules-consul-acl/compare/0.0.1...0.1.0